+++ backoffice/Domain/Entidades/Cadastros/Gerador/Gerador.cs
@@ -9,9 +9,9 @@ namespace Domain.Entidades.Cadastros.Gerador
         [Key]
         public int Id { get; set; }
         public string NomeGerador { get; set; }
-        public int QuantidadeHoras { get; set; }
+        public long QuantidadeHoras { get; set; }
         public DateTime DataUltimaTrocaOleo { get; set; }
-        public int QuantidadeHorasTroca { get; set; }
+        public long QuantidadeHorasTroca { get; set; }
         [ForeignKey("Empresa")]
         public int IdEmpresa { get; set; }
